Make src/ folder and move things there
Saw this on the #BoringShow and decided to follow it.
